avatar posicionado -siguiente cambios en estilo infomodal
