+++ src/front/components/InfoModalUser.jsx
@@ -3,24 +3,13 @@ import React, { useEffect, useState } from "react";
 import Timer from "./Timer";
 import GameContainer from "../pages/GameContainer";
 import "../CSS/Game.css";
-import { Dropdown, Button, OverlayTrigger, Tooltip } from 'react-bootstrap'
-import Avatar_01 from '../assets/img/UI/Avatars/Avatar_01.png';
-import Avatar_02 from '../assets/img/UI/Avatars/Avatar_02.png';
-import Avatar_03 from '../assets/img/UI/Avatars/Avatar_03.png';
-import Default_Avatar from '../assets/img/UI/Avatars/default_avatar.png';
+import { OverlayTrigger, Tooltip } from 'react-bootstrap'
 
-const avatarMap = {
-  "Avatar_01.png": Avatar_01,
-  "Avatar_02.png": Avatar_02,
-  "Avatar_03.png": Avatar_03,
-  "default_avatar.png": Default_Avatar,
-};
-
-export const InfoModalUser = ({ showEnigma }) => {
+export const InfoModalUser = ({ showEnigma, isOpen, onClose }) => {
 
   const {user, nivelActual, totalHintsUsed, isUserLoading  } = useGame()
 
-  if (showEnigma) {
+  if (showEnigma || !isOpen) {
     return null;
   }
 
@@ -48,23 +37,13 @@ export const InfoModalUser = ({ showEnigma }) => {
 
   const hints = [1, 2, 3];
 
-const avatarSrc = user.avatar_filename 
-    ? avatarMap[user.avatar_filename] || Default_Avatar 
-    : Default_Avatar; 
-
   return (
 
-<div className="PlayerMenu-toggle">
-    <Dropdown drop="end">
-      <Dropdown.Toggle className="Avatar-Button" as="div">
-                <img src={avatarSrc} className="Avatar" alt="Avatar"/> 
-      </Dropdown.Toggle>
-
-      <Dropdown.Menu className="p-3">
+<div className="info-modal-user-content">
+<button onClick={onClose} className="close-modal-button">X</button>
         <h2>Informe Clasificado</h2>
 
         <div className="d-flex align-items-center gap-3 mb-3">
-      
           <h3 className="modalUser-name">{user.username}</h3>
         </div>
 
@@ -97,8 +76,6 @@ const avatarSrc = user.avatar_filename
             );
           })}
         </div>
-      </Dropdown.Menu>
-    </Dropdown>
 </div>
   );
 };
